Add adddel.js file and edit posibility
