+++ webpack.config.js
@@ -6,6 +6,7 @@ module.exports = {
   entry: {
     index: './src/index.js',
     print: './src/print.js',
+    adddel: './src/adddel.js',
   },
   devServer: {
     static: './dist',
